requirements.txt
mysql-connector-python==8.1.0
numpy==1.25.2
pandas==2.0.3
pyodbc==5.0.1
